third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-prefixed-after-unprefixed-expected.txt


EVENT(loadstart)
Call prefixed EME API after unprefixed EME API.
RUN(mediaKeys = new MediaKeys("org.w3.clearkey"))
EXPECTED (mediaKeys != 'null') OK
RUN(video.setMediaKeys(mediaKeys))
TEST(video.webkitGenerateKeyRequest('webkit-org.w3.clearkey')) THROWS(DOMException.INVALID_STATE_ERR: Failed to execute 'webkitGenerateKeyRequest' on 'HTMLMediaElement': Mixed use of EME prefixed and unprefixed API not allowed.) OK
END OF TEST
